somem cleanup
